feat: Phase 4 - NoteEditorFragment を削除し Activity が setContent を直接呼ぶ形に移行
- NoteEditorActivity を AppCompatActivity + setContent に書き換え
  - ActivityResultLauncher 群を Fragment から Activity に移動
  - EmojiSelection インターフェースを Activity が実装
  - サイドエフェクト（アカウント切り替え/認証チェック/ファイルサイズ警告）を
    setupSideEffects() に集約
  - DataBinding / ViewCompat.setOnApplyWindowInsetsListener を削除
- NoteEditorFragment.kt を削除
- activity_note_editor.xml を削除
- NoteEditorScreen: contentWindowInsets = WindowInsets.safeDrawing に戻す
- NoteEditorScreen: bottomBar に imePadding() を追加（キーボード表示時対応）
- BUILD SUCCESSFUL 確認済み
