+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/editor/NoteEditorScreen.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeDrawing
 import androidx.compose.foundation.rememberScrollState
@@ -139,7 +140,7 @@ fun NoteEditorScreen(
             }
         },
         bottomBar = {
-            Column {
+            Column(modifier = Modifier.imePadding()) {
                 NoteFilePreview(
                     noteEditorViewModel = viewModel,
                     onShow = { feedback.performClickHapticFeedback(); onShowMediaPreview(it) },
