--- modules/features/note/src/main/res/layout/activity_note_editor.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools">
-    <data>
-        <import type="android.view.View"/>
-        <import type="net.pantasystem.milktea.common_android.ui.SafeUnbox" />
-    </data>
-    <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            tools:context="net.pantasystem.milktea.note.NoteEditorActivity">
-
-       <FrameLayout
-               android:id="@+id/fragmentBase"
-               android:layout_width="match_parent"
-               android:layout_height="match_parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
